wip: update docs, add more tests
